Revert "Revert "New horizontal layout and font change.""
This reverts commit f8e9b3626991e51da8a048bb17987959dfb5b5cf.

The revert that this reverts was an intentional revert of an
accidental merge. However, because it was a revert of pulling
in master, it effected itself as a delete of added files. Then
these changes were merged into master, deleting those files from
master. Because those changes from master had been merged and then
reverted, nothing complained about a lack of fast forward, etc.